fix: make omega audit runnable via node command
